Fix fragmentary-field-reading logic in drizzle_state_field_read() so that it doesn't accidentally cross a packet boundary. Adjusted types of some field variables to match the largest values that can be stored in them.

+++ libdrizzle/result.h
@@ -64,14 +64,14 @@ struct drizzle_result_st
   uint64_t row_count;
   uint64_t row_current;
 
-  uint16_t field_current;
-  size_t field_total;
-  size_t field_offset;
-  size_t field_size;
+  uint16_t field_current;         /* field index */
+  uint64_t field_total;           /* total length of the field currently being read */
+  uint64_t field_offset;          /* offset within field of most recently read field fragment (0 if first/only fragment) */
+  uint32_t field_size;            /* size of most recently read field value or fragment of field value; max 2^24 */
   drizzle_field_t field;
   drizzle_field_t field_buffer;
 
-  uint64_t row_list_size;
+  size_t row_list_size;
   drizzle_row_t row;
   drizzle_row_t *row_list;
   size_t *field_sizes;
